+++ deploy/configs/ppdataaug_ocr_img2img.yaml
