+++ tools/predict.py
@@ -20,7 +20,7 @@ sys.path.append(os.path.abspath(os.path.join(__dir__, '../deploy/')))
 from ppcv.core.config import ArgsParser
 
 from python.ppldi import PPDataImprove
-from python.ppeda import PPEasyDataAug
+from python.ppdataaug import PPDataAug
 from utils.utils import load_yaml
 
 def argsparser():
@@ -65,7 +65,7 @@ if __name__ == '__main__':
         ppldi = PPDataImprove(args)
         ppldi.run()
     elif "DataGen" in yaml_data:
-        ppeda = PPEasyDataAug(args)
-        ppeda.run()
+        ppdataaug = PPDataAug(args)
+        ppdataaug.run()
     else:
         raise Exception("Error config")
